chore(tests): refactoring
