--- packages/ui-kit/src/components/Counter/__test__/mockData.ts
@@ -1,3 +1,0 @@
-export const counter = {
-  value: '123'
-}
