--- packages/ui-kit/src/components/Leaderboard/__test__/mockData.ts
@@ -1,8 +1,0 @@
-export const leaderboard = {
-  headers: ['header-1', 'header-2', 'header-3'],
-  rows: [
-    ['dim-1', 'dim-2', '30'],
-    ['dim1', 'dim-2', '60'],
-    ['dim-1', 'dim-2', '90']
-  ]
-}
