+++ packages/ui-kit/src/components/TimeSeries/TimeSeries.test.tsx
@@ -1,20 +1,52 @@
 import { render } from '@testing-library/react'
 import { Chart } from 'chart.js'
 import React from 'react'
-import { RelativeTimeRange, TimeSeriesGranularity } from '../../../helpers'
+import { mockTimeSeriesQuery, RelativeTimeRange, TimeSeriesGranularity } from '../../helpers'
+import { Dom, setupTestHandlers } from '../../helpers/testing'
 import { TimeSeries } from '../index'
-import { timeSeries } from './mockData'
-import { setupHandlers } from './mswHandlers'
+
+const mockData = {
+  labels: ['2023-07-01', '2023-07-02', '2023-07-03'],
+  values: ['10', '20', '15']
+}
+
+// @TODO: why the values type for static data is different from the one from the server?
+const mockStaticData = {
+  labels: ['2023-07-01', '2023-07-02', '2023-07-03'],
+  values: [10, 20, 15]
+}
+
+const handlers = [
+  mockTimeSeriesQuery((req, res, ctx) => {
+    const { metricName } = req.variables.timeSeriesInput
+
+    if (metricName === 'should-fail') {
+      return res(
+        ctx.errors([
+          {
+            message: 'Something went wrong'
+          }
+        ])
+      )
+    }
+
+    return res(
+      ctx.data({
+        timeSeries: mockData
+      })
+    )
+  })
+]
 
 describe('TimeSeries', () => {
-  let dom: ReturnType<typeof render>
+  let dom: Dom
 
   beforeEach(() => {
-    setupHandlers()
+    setupTestHandlers(handlers)
   })
 
   it('should render static data', () => {
-    dom = render(<TimeSeries labels={timeSeries.labels} values={timeSeries.values} />)
+    dom = render(<TimeSeries labels={mockStaticData.labels} values={mockStaticData.values} />)
 
     const chartElement = dom.getByRole('img') as HTMLCanvasElement
     const chartInstance = Chart.getChart(chartElement)
@@ -22,8 +54,8 @@ describe('TimeSeries', () => {
     const chartData = chartInstance?.data.datasets[0].data
     const chartLabels = chartInstance?.data.labels
 
-    expect(chartLabels).toEqual(timeSeries.labels)
-    expect(chartData).toEqual(timeSeries.values)
+    expect(chartLabels).toEqual(mockStaticData.labels)
+    expect(chartData).toEqual(mockStaticData.values)
   })
 
   it('should render data from server', async () => {
@@ -47,11 +79,12 @@ describe('TimeSeries', () => {
     const chartData = chartInstance?.data.datasets[0].data
     const chartLabels = chartInstance?.data.labels
 
-    expect(chartLabels).toEqual(timeSeries.labels)
-    expect(chartData).toEqual(timeSeries.values.map((value) => Number(value)))
+    expect(chartLabels).toEqual(mockData.labels)
+    expect(chartData).toEqual(mockData.values.map((value) => Number(value)))
   })
 
   it('should show error fallback when query fails', async () => {
+    console.error = jest.fn()
     dom = render(
       <TimeSeries
         query={{
@@ -68,8 +101,9 @@ describe('TimeSeries', () => {
     )
 
     await dom.findByText('Unable to connect')
-
     await dom.findByText('Sorry we are not able to connect at this time due to a technical error.')
+
+    expect(console.error).toHaveBeenCalled()
   })
 
   it('should show error fallback on props mismatch', async () => {
