--- packages/ui-kit/src/components/TimeSeries/__test__/mockData.ts
@@ -1,4 +1,0 @@
-export const timeSeries = {
-  labels: ['2023-07-01', '2023-07-02', '2023-07-03'],
-  values: ['10', '20', '15']
-}
